requirements-client.txt
# Telegram Client API
telethon==1.34.0

# Async support
aiohttp==3.9.1

# Configuration
pydantic==2.5.3
PyYAML==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Scheduling
APScheduler==3.10.4

# Utilities
cryptography==41.0.7
